Split out arrow-string (#2594)
