+++ .github/workflows/arrow_flight.yml
@@ -31,10 +31,11 @@ on:
       - arrow-buffer/**
       - arrow-cast/**
       - arrow-data/**
-      - arrow-schema/**
-      - arrow-select/**
       - arrow-flight/**
       - arrow-ipc/**
+      - arrow-schema/**
+      - arrow-select/**
+      - arrow-string/**
       - .github/**
 
 jobs:
