+++ .github/workflows/integration.yml
@@ -24,20 +24,21 @@ on:
       - master
   pull_request:
     paths:
-      - arrow/**
+      - .github/**
       - arrow-array/**
       - arrow-buffer/**
       - arrow-cast/**
-      - arrow-data/**
-      - arrow-schema/**
-      - arrow-select/**
-      - arrow-ipc/**
       - arrow-csv/**
-      - arrow-json/**
-      - arrow-pyarrow-integration-testing/**
+      - arrow-data/**
       - arrow-integration-test/**
       - arrow-integration-testing/**
-      - .github/**
+      - arrow-ipc/**
+      - arrow-json/**
+      - arrow-pyarrow-integration-testing/**
+      - arrow-schema/**
+      - arrow-select/**
+      - arrow-string/**
+      - arrow/**
 
 jobs:
