+++ Cargo.toml
@@ -17,23 +17,24 @@
 
 [workspace]
 members = [
-        "arrow",
-        "arrow-array",
-        "arrow-buffer",
-        "arrow-cast",
-        "arrow-csv",
-        "arrow-data",
-        "arrow-flight",
-        "arrow-integration-test",
-        "arrow-integration-testing",
-        "arrow-ipc",
-        "arrow-json",
-        "arrow-schema",
-        "arrow-select",
-        "object_store",
-        "parquet",
-        "parquet_derive",
-        "parquet_derive_test",
+    "arrow",
+    "arrow-array",
+    "arrow-buffer",
+    "arrow-cast",
+    "arrow-csv",
+    "arrow-data",
+    "arrow-flight",
+    "arrow-integration-test",
+    "arrow-integration-testing",
+    "arrow-ipc",
+    "arrow-json",
+    "arrow-schema",
+    "arrow-select",
+    "arrow-string",
+    "object_store",
+    "parquet",
+    "parquet_derive",
+    "parquet_derive_test",
 ]
 # Enable the version 2 feature resolver, which avoids unifying features for targets that are not being built
 #
