+++ arrow/Cargo.toml
@@ -54,12 +54,12 @@ arrow-ipc = { version = "28.0.0", path = "../arrow-ipc", optional = true }
 arrow-json = { version = "28.0.0", path = "../arrow-json", optional = true }
 arrow-schema = { version = "28.0.0", path = "../arrow-schema" }
 arrow-select = { version = "28.0.0", path = "../arrow-select" }
+arrow-string = { version = "28.0.0", path = "../arrow-string" }
 rand = { version = "0.8", default-features = false, features = ["std", "std_rng"], optional = true }
 num = { version = "0.4", default-features = false, features = ["std"] }
 half = { version = "2.1", default-features = false, features = ["num-traits"] }
 hashbrown = { version = "0.13", default-features = false }
 regex = { version = "1.7.0", default-features = false, features = ["std", "unicode", "perf"] }
-regex-syntax = { version = "0.6.27", default-features = false, features = ["unicode"] }
 packed_simd = { version = "0.3", default-features = false, optional = true, package = "packed_simd_2" }
 chrono = { version = "0.4.23", default-features = false, features = ["clock"] }
 comfy-table = { version = "6.0", optional = true, default-features = false }
@@ -92,7 +92,7 @@ force_validate = ["arrow-data/force_validate"]
 ffi = ["bitflags"]
 # Enable dyn-comparison of dictionary arrays with other arrays
 # Note: this does not impact comparison against scalars
-dyn_cmp_dict = []
+dyn_cmp_dict = ["arrow-string/dyn_cmp_dict"]
 # Enable dyn-arithmetic kernels for dictionary arrays
 # Note: this does not impact arithmetic with scalars
 dyn_arith_dict = []
